Fix tests

+++ python/ray/ml/tests/test_tensorflow_predictor.py
@@ -20,7 +20,6 @@ def build_model() -> tf.keras.Model:
             tf.keras.layers.Dense(1),
         ]
     )
-
     return model
 
 
@@ -53,7 +52,7 @@ def test_predict_array_with_preprocessor():
     predictions = predictor.predict(data_batch)
 
     assert len(predictions) == 3
-    assert predictions.to_numpy().flatten().round().tolist() == [2, 4, 6]
+    assert predictions.to_numpy().astype(float).flatten().round().tolist() == [2, 4, 6]
     assert hasattr(predictor.preprocessor, "_batch_transformed")
 
 
@@ -67,4 +66,4 @@ def test_predict_array():
     predictions = predictor.predict(data_batch)
 
     assert len(predictions) == 3
-    assert predictions.to_numpy().flatten().tolist() == [1, 2, 3]
+    assert predictions.to_numpy().astype(float).flatten().tolist() == [1, 2, 3]
